core: DiskValidator#isDiskAttachedToAnyVm streams
This patches changes the implementation of the isDiskAttachedToAnyVm
method to first generate the list of offending VM names (which should
be a non-op if there are none), and then perform the check on the
generated string instead of the original list.

While this patch doesn't have too much meaning on its own, it sets up
the subsequent patch, which will add another filter clause to the
aforementioned stream.

Change-Id: I7c89929c040c9e6cb7913b24ba2484db34562752
Signed-off-by: Allon Mureinik <amureini@redhat.com>