core: Allow transferring images for down VMs
Currently, DiskValidator#isDiskAttachedToAnyVm() is used to block
image transfers on all non-floating disks.

This behavior seems to be a bit too naive, as there's no
storage-related not to upload/download an image that happens to have
a database entry that ties it to a VM.

For simplicity's sake, the operation is still blocked for any status
other than DOWN. In the future, we can consider refining this
restriction.

Change-Id: Ia313a451f7642b10b1ee3a46d2ce1bdd1d2fa061
Bug-Url: https://bugzilla.redhat.com/1414499
Signed-off-by: Allon Mureinik <amureini@redhat.com>