webadmin: image upload - remove fakepath prefix
Removed 'C:\fakepath\'[1] prefix added by the browser
to the selected file name.

[1] According to the specifications of HTML5,
a file upload control should not reveal the real local
path to the file you have selected, if you manipulate
its value string with JavaScript. Instead, the string
that is returned by the script, which handles the file
information is c:\fakepath.

Source:
https://acidmartin.wordpress.com/2009/06/09/the-mystery-of-cfakepath-unveiled/

Change-Id: Ia873b88f322a2ded905744753ab6134596bf45c7
Bug-Url: https://bugzilla.redhat.com/1461932
Signed-off-by: Daniel Erez <derez@redhat.com>