backend: remove character restrictions from network names
Currently, oVirt limits the names of its networks to 15
alphanumeric characters. This limitation dates back to
the ages where all networks were VM network, all were
implemented by a Linux bridge, and the same name was
used to identify the network and the Linux bridge
implementing it on each host.

That has to change. 15 characters are not enough for
humans; spaces, and other special characters are
visually useful, and non-English speaking users
would love to use their native alphabet in network names.

For this reason network names are allowed to consist of 256 Unicode
characters. If a Linux bridge is created, the name is created
from the network name, if it is suitable, or from the network's uuid.

(cherry-picked from commit 0cee2277bbbf176a6ed0f56e40a04ff56b86f1f5)

Bug-Id: https://bugzilla.redhat.com/show_bug.cgi?id=1476823
Change-Id: Idadf272959ae10afc1de7ed6ce6daed70b78b73d
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>