backend: improve problematic block domains sync
When a block storage domain's status is StorageDomainStatus.Inactive but
vdsm reports that it is DomainMonitoringResult.OK, the domain's status
is switched to StorageDomainStatus.Active.

Today, the engine calls SyncLunsInfoForBlockStorageDomain for each such
domain in a new thread using the same host.
This patch improves this behavior by synchronizing all the domains
together using SyncStorageDomainsLuns and thus performing one call to
vdsm (instead one per domain) in one thread.

Change-Id: I1d8ddb18e5b40006c88eb632801920c14045db71
Bug-Url: https://bugzilla.redhat.com/1468974
Signed-off-by: Idan Shaby <ishaby@redhat.com>