restapi: support setting empty CCV value for a VM via REST api
Trying to set the Custom Compatibility Version of a VM to an empty/null
value via REST api (in order to take the Custom Compatibility Version value
of the Cluster) was not supported/handled.

This patch fixed it, so that for setting the Custom Compatibility Version
of a VM to null, the REST api request body should include:
<vm>
    <custom_compatibility_version/>
</vm>

This will be translated to sending entity.customCompatibilityVersion set
to null as required.

Change-Id: I6cdc527b3fca98e0ee230275563539fd28020966
Bug-Url: https://bugzilla.redhat.com/1464819
Signed-off-by: Sharon Gratch <sgratch@redhat.com>