scheduling: Hide InClusterUpgrade cluster policy
The cluster policy was written to workaround a bug during
EL6 to EL7 Host OS upgrades. We no longer support EL6
so there is no need to show it to users.

It can still be recreated manually if an use case shows
up, because the necessary policy unit is still available.

Bug-Url: https://bugzilla.redhat.com/1469478
Change-Id: I846e44a485595699e703dfd3c08c796d29169763
Signed-off-by: Martin Sivak <msivak@redhat.com>