db: move permission data changes to SP call
Postgres 9.3 introduced Materialized Views support.
In order to use that in oVirt we need to find out which views are
relatively complex and have a low data change rate.

The *permissions* views are a natural candidate for that, but before
that we need to change all direct changes on the permission table to use
a SP call in order to refresh the materialized views in one place.

This patch changes all data changes to use a SP call.

Change-Id: I4fb7cbdae37083b60938195256257cf848f133c0
Signed-off-by: emesika <emesika@redhat.com>