core: set new Pool's images format to COW for 'auto select target'
In case new pool is created based on a template with disk(s) and the
'Auto select target' checkbox (in 'Resource allocation' tab) is set,
all Pool VM's images are required to be created with volume format set to
the same format as the template's disk (RAW or COW).
If RAW format is required for the Pool VM's images, then the Pool
creation failed on validation phase.

Because of that problem, the 'auto select target' feature isn't working and it is a
regression (the regression was added in change id: I6f0f143).

This patch fixes it by setting each diskImage.volumeFormat of the Pool's template
to COW (or RAW if the disk storage type is CINDER) before creating the Pool's VMs.

Change-Id: I9fd04cbcea08125e06c0b9ffb237bcb53057a61a
Bug-Url: https://bugzilla.redhat.com/1468301
Signed-off-by: Sharon Gratch <sgratch@redhat.com>