core: Hide expected network exceptions during host installation
those timeout exceptions are expected, thus in this case are actually
not exceptions from point of view of our system.

as a side effect PollVDSCommand won't log VDSNetworkException as a part
of processing catch block of TimeoutException. Currently it should not
be problem, since Poll is called only from HostSetupNetworkPoller,
which in turn is only called from:

• HostSetupNetworksCommand (because of this command we do this change)
• HostConnectivityChecker (should not be problem, this command
requires only boolean yes/no result, and causing TimeoutException
will be logged.)

Change-Id: I6dbf21e7b9d00ec86a1c646fe8dc47755eab5644
Bug-Url: https://bugzilla.redhat.com/1442687
Signed-off-by: Martin Mucha <mmucha@redhat.com>