core: DiskValidator#isDiskPluggedToAnyNonDownVm(true)
Commit 8d3cc2 unified the functionality of
isDiskPluggedToAnyNonDownVm and isDiskPluggedToVmsThatAreNotDown, but
botched the condition where checkOnlyVmsSnapshotPluggedTo=true.

In case it was set to true, the original
isDiskPluggedToVmsThatAreNotDown skipped VMs that had snapshots
plugged to them, while the unified code includes them in the filter
instead of excluding them.

This patch rectifies that mistake, and adds a proper test to assure
it is not broken again.

Change-Id: Ib7c02f8f621b98f252e8d81bcf3852c5a8d8fdf6
Signed-off-by: Allon Mureinik <amureini@redhat.com>