webadmin: make hosts table status icon column wider
When the alert icon is displayed in the status icon column of
the host table, it was getting cut off. The column width has
been increased to allow it to be seen with the standard icons.

Status icons wider then 19px (currently only the error status)
will still cause the alert icon to be only partially rendered.

Change-Id: I37474f9ded7cfd25222070216ae9d031b411e1ec
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>