backend: direct LUN should not have a PV ID in the DB
When a lun is used as a part of a storage domain, it contains a PV ID in
the luns table.
However, when it is used as a direct LUN, vdsm doesn't create a PV on it
and thus it shouldn't contain a PV ID in that table.

This patch makes sure that a lun won't have an old PV ID in the DB (the
PV ID of that lun which was previously a part of a storage domain) in
case it is used purely as a direct LUN, and not both as a part of a
storage domain and as a direct LUN.

Change-Id: I427bfae9102e1465d7f00d18bdf9cf9372b6c1a1
Bug-Url: https://bugzilla.redhat.com/1475767
Signed-off-by: Idan Shaby <ishaby@redhat.com>