core: robust handling of the top-level element in ovf
The OVF specification indicates that the top-level element of an OVF
should be "Envelope" but doesn't state whether it should be with an
XML namespace or not. In some examples the element is "ovf:Envelope"
and in other example it is simply "Envelope".
This patch changes the OVF reader to be able to parse both cases.

Change-Id: I2ec7259ecfaf5c48b61d8e580a5f1031b6c45eeb
Signed-off-by: Arik Hadas <ahadas@redhat.com>