backend: don't remove VG ID of direct lun that is a part of a SD
When synchronizing a direct lun, the VG ID that comes from getDeviceList
used to be changed to an empty string before updating the lun in the DB,
assuming that a direct lun doesn't have a VG ID and thus the
information that comes from vdsm is old and irrelevant.

That isn't true when talking about a lun which is a part of a storage
domain (and thus has a VG ID) and is also used as a direct lun.

Therefore, this patch makes sure that when synchronizing a direct lun,
the VG ID will be updated only if it is not an empty string.

Change-Id: I3171bee8bc0d45c459977638919489bb25cc4e49
Bug-Url: https://bugzilla.redhat.com/1471759
Signed-off-by: Idan Shaby <ishaby@redhat.com>