core: proper NetworkSection in OVF
The OVF specification indicates that the networks section should start
with NetworkSection element rather than Section element with an attribute
of type=NetworkSection_Type. It should also include the logical networks
used by the VM/template instead of a single network called "Network 1"
that we generated until now.

As for backward compatibility, this section was never read by oVirt so
we are safe.

Change-Id: I022a768dbf1bc948cb97eacfb94389f6ecfe9026
Signed-off-by: Arik Hadas <ahadas@redhat.com>