core: better alignment with ovf specification for disks
The "File" elements in the References section are designed to enable tools
to verify the integrity of the OVF package (OVA). They are supposed to be
minimal. However, we abuse them by adding additional information about the
disk in these elements.

So in order to be better aligned with the specification, this patch changes
the OVF reader to consume that data from the "Disk" elements inside the
"DiskSection" part, without breaking backward compatibility.

Change-Id: Idcbc86decd53d2495b5f8ece78acf52cbef60bc4
Signed-off-by: Arik Hadas <ahadas@redhat.com>