core: remove redundant cast from correlation_id
In the search engine, the ORDER BY  includes a cast to varchar of
correlation_id (correlation_id::varchar), from the other hand, the
selection list has only correlation_id.

Since a column name that appears in ORDER BY must be included exactly in
the same format in the selection list (correlation_id::varchar and
correlation_id are not considered the same), the query fails

This patch removes the redundant casting of correlation_id to varchar
and makes it possible to sort Events by correlation_id.

Change-Id: I440608e7858911ea0df37629475fe2e6d97913f7
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1471815
Signed-off-by: emesika <emesika@redhat.com>