webadmin: add 'optimized for' field to vm/template/pool/v2v general sub tab
Since we added a new "optimized for" field type of HP VMs to the VM dialog
(in addition to the Desktop/Server types), we also want to display the current
chosen "optimized for" type for a VM/Template/Pool (in addition to the
VM type icon implemented only for VMs).

For that reason, this patch added the "optimized For" field to the
General sub tab for VM, Pool, Template and imported VMs (on import VMs
dialog).

Change-Id: Ic65223e8d6cf86f7cb43ea060b0fe22d7fe3fbd9
Bug-Url: https://bugzilla.redhat.com/1457239
Signed-off-by: Sharon Gratch <sgratch@redhat.com>