core: use ovf:href instead of ovf:fileRef to identify disk id
According to the OVF specification, ovf:fileRef attribute of a disk should
point to the ovf:id attribute of a File element within the References
section. However, we used that attribute exactly as the ovf:href attribute
of a File element so it was also used to retrieve the disk's (image group's)
id. This patches changes that so ovf:href will be used instead.
We should note that we still can't change the content of ovf:fileRef to
something else than the relative path of the volume because previous
versions still count on that.

Change-Id: I2b3a8f379a68ad089547f85d8d93d2bbb5213c54
Signed-off-by: Arik Hadas <ahadas@redhat.com>