core: Added invalidate cache mechanism for domain files refresh
Unless force refresh was specified every call to list a domain files was
returning cached results from the DB rather than getting the files list
from the storage, this was changed to invalidate the cache after a
specified number of minutes from the last refresh (default is 1 minute and
can be changed by changing the RepoDomainInvalidateCacheTimeInMinutes
configuration value).

The old mechanism was based on a periodical refresh every 60 minutes
(configurable) regardless of whether there was a request for the files
list, this mechanism was removed.

Change-Id: I8eccea6f440cf0050626f95a1999af912c575287
Bug-Url: https://bugzilla.redhat.com/1414207
Signed-off-by: Tal Nisan <tnisan@redhat.com>