core: Explicitly require JEE concurrency API
A recent patch changed the 'utils' module to use the JEE concurrency
API, but it didn't explicitly include the required artifact in the POM
file. In some environemts this means that the project won't build
correctly, in particular it doesn't build correctly in the latest
version of IntelliJ IDEA. This patch adds the dependency, as it is good
practice and solves the issue in this environment.

Change-Id: Ib449fcd2d72c8ed3d7d3f6964264a94931e2d990
Signed-off-by: Juan Hernandez <juan.hernandez@redhat.com>