core: Hard code session IDs in tests
This patch adds engine session IDs to the fixtures file so that the
DAO tests can use them.
This removes the dependency that QuotaDaoTest and PermissionDaoTest
currently have on EngineSessionDao, which in theory is unrelated to
them and can cause false positives in case there's a problem with it.
In other words, this patch improves the isolation of the DAO tests.

Another improvement gained here is the removal of two DAO calls from
each of these tests' setup() methods, which should improve their
speed.

Change-Id: Ie85830335420f68e4b60ad936359af0213dc4c6f
Signed-off-by: Allon Mureinik <amureini@redhat.com>