webadmin: resize dashboard on window resize
Previously, the dashboard (and other ui plugins) had their
height calculated at iframe load time. However, when a user
resized the window, the iframe height was not recalculated,
often leading to a duplicate vertical scrollbar (one for the
page, one for the iframe.)

Fixed by simply calling the frame resize handler on page
resize as well.

Change-Id: I3af0f64bf4d7013dd3d31f675490368e5bdd7c7c
Signed-off-by: Greg Sheremeta <gshereme@redhat.com>