core: prevent NPE on change of rng of a running vm
We try to preserve the RNG device rather than introducing a new one in
case one changes the rng device of a running VM. However, there was no
protection for the case that someone asks to add an RNG device for a
running VM, and that results in NPE (that was detected by Coverity's
static code analysis). This patch introduces a fix for that - in that
case, we need to add a new RNG device.

Change-Id: I80a231ca8c02890bb9ca7340b8a59973c3c759eb
Signed-off-by: Arik Hadas <ahadas@redhat.com>