core: handle HE VM hosting host fencing
The non-responding host is in a fencing flow while the engine VM is
restarted on another machine, the fencing flow did only the <stop>
step from the fencing flow and then the engine VM goes down.
Since in the engine start we look only for host with <non-responding> status ,
this host which is actually in <down> status will not be started.

Proposed solution is to add a is_in_fencing_flow boolean column to vds_dynamic

1) The flag is set at the start of hard fencing flow (stop command) and reset
   at the end of the hard fencing flow (start command) when the origin
   of those command comes from non-responding host and not manual fencing

2) In the non-responding treatment that occurs on engine start we will take
   in account also hosts that have the is_in_fencing_flow
   and reset the is_in_fencing_flow flag

3) when the host is UP either by a successful start command or by Host Monitoring,
   the is_in_fencing_flow should be rest as well

This will ensure that engine will try to fence hosts that were stuck in the middle
of a fencing flow as a result of engine restart

Change-Id: I8761b52b5ffb4e10875a5b235a4cfc8de6664cf9
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1440101
Signed-off-by: emesika <emesika@redhat.com>