webadmin: refactor configure popup, tab contents and dialog tab panel
For the configuration popup, refactor the code to put as much view
setup in the UiBinder file as possible.  Each content tab was also
refactored to layout the content in bootstrap/PatternFly containers
and to put as much view setup in UiBinders as possible.  All config
tab content has also been configured to scroll vertically as needed.

Updated dialog tab panel such that a tab's content has no horizontal
margin allowing the tab's content to use the entire space or use the
default bootstrap/PatternFly container margins.

For the SharedMaxPoolView, refactored the layout code into a UIBinder
to maintain consistency between all of the config popup tab widgets.

Change-Id: I60bb1729d4b8297d69f710debc8d130f8a8bb53e
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>