core: Remove storage pool from StoragePoolIsoMapDaoTest
The only aspect of the storage pool used in StoragePoolIsoMapDaoTest
is it's ID, which is just given as a constant in FixturesTool.

This patch removes the redundant objects from the
StoragePoolIsoMapDaoTest to make it clearer and easier to maintain.
As a side bonus, it also removes another DAO call, speeding up the
test.

Change-Id: I6f214062fbd419547eac5a65e2b15b88af4b887c
Signed-off-by: Allon Mureinik <amureini@redhat.com>