core: Remove StorageServerConnection from StorageServerConnectionLunMapDaoTest
The only part of the StorageServerConnection used in
StorageServerConnectionLunMapDaoTest is its ID, which can just be
taken from the FixturesTool constant.

This patch removes the roundtrip to the database that retrieves the
unneeded object in order to clean up the test and make it easier to
maintain. As a side bonus, removing this DAO call from the setup also
speeds up the test.

Change-Id: Ib7a62010021111069604c9b9cf5f01fc2ce891c3
Signed-off-by: Allon Mureinik <amureini@redhat.com>