core: prevent NPE when exporting networkless NIC
As part of the recent changes aim to align the OVF generation in oVirt
with the standard specification of OVF, we changed the OVF generator
to write the actual network names in the NetworkSection rather than a
single network named "Network 1". What we missed, however, is that a
NIC may not be associated with any network. This leads to an NPE in
the OVF generation process. Adding a null-check to prevent that.

Change-Id: I7d120cfaf75f5c0036b3df0062fed2600945e4da
Bug-Url: https://bugzilla.redhat.com/1476191
Signed-off-by: Arik Hadas <ahadas@redhat.com>