core: do not use placeholder for nic's network
This place-holder just complicates things. We can just generate
the NIC's part in the domain XML that fits for legacy bridges and
let VDSM replace the inner 'source' element in case it is a non
legacy bridge. The innder 'source' element that VDSM is provided
with contains the network name and that is the only information
it apparently needs.

Change-Id: I377c52fde263dcf8f8318d180884185574365d6a
Signed-off-by: Arik Hadas <ahadas@redhat.com>