webadmin: Use default grid item comparator
If a grid has no explicit column sorting applied by the user
(which is the initial state of all grids), use a default item
comparator with following logic:

- first, try to compare by name (lexo-numeric) if the item is
  assignable to Nameable interface (this covers most, but not
  all, business entities), or treat it as null

- otherwise (for items still considered equal by above rule),
  try to compare by queryable ID if the item is assignable to
  Queryable interface (this should cover the rest of entities),
  or treat it as null

- finally, put "null" items at the end (nulls last strategy)

Above comparator logic is in-line with the general UI design
of having "Name" column with clickable links leading to detail
views for each main grid.

Above comparator is not applied if a grid has been explicitly
sorted by the user (by clicking the given column header).

Change-Id: I4711af76410a16354c38c083130be1c2e033306f
Bug-Url: https://bugzilla.redhat.com/1445235
Signed-off-by: Vojtech Szocs <vszocs@redhat.com>