core: Display an alert if the VM has illegal images
Display an alert next to the VM if the VM has illegal images. A VM may
have illegal images if, for example, there was a failure during merge.
It is important to display an alert because an illegal image may
indicate that the volume on the storage is illegal, because if the
VM is stopped, the user will not be able to start it again before fixing
the illegal volume.

Change-Id: I46bdceea155774bbb7641dd767c3caf4f53b8102
Signed-off-by: Ala Hino <ahino@redhat.com>
Signed-off-by: Allon Mureinik <amureini@redhat.com>