core: undo recent generalization of ovf processing
Reverting the attempt to generalize OvfReader and OvfWriter to support
different kinds of OVF implementations. It produces a complicated code
and increases the chance for regressions. A better approach would be
to split the OVF handling of ordinary OVFs in oVirt and general OVFs
that better comply with the specification.

Change-Id: Ibbe176cfc0a5bcb0319d178f573ba948ab1d6fc8
Signed-off-by: Arik Hadas <ahadas@redhat.com>