core: separate out ovirt-specific part in ovf writing
There are several parts in the OVF writing in oVirt that don't comply
with the OVF specification. This patch separates out those part in OVF
writing from the OvfWriter class to OvfOvirtWriter. The writer that is
going to be added for writing OVFs that are better aligned with the OVF
specification will also extend OvfWriter.

Change-Id: If775da94c8570264591b42a18214a8cf918e03c4
Signed-off-by: Arik Hadas <ahadas@redhat.com>