core: move common ovf handling up in the chain
We had the following pattern for writing applications list and mac
address to OVFs:
- abstract methods in OvfWriter
- empty implementations for templates
- full implementations for VMs

This pattern creates a problem when extending the chain of OVF writers
(it needs to be extended with OVF writers to OVAs). Therefore, this
pattern is replaced in this patch with the following pattern:
- full implementations in OvfWriter
- empty implementations for templates

This pattern fits as long as the same handling should be done for the
different kinds of OVFs we are going to produce.

Change-Id: I3ffd2c806f89f0fc0238df84b95140b32233278f
Signed-off-by: Arik Hadas <ahadas@redhat.com>