core: Verify that the VM exists when returning callback
It seems that there are use cases where a merge operation fails, leaving
a step in ACTIVE state in the command_entities table. However, at
some point, the VM is deleted. In this case, i.e. as long as the failed
step is ACTIVE, engine restart fails because we will be looking for a VM
that doesn't exist anymore and failing on NPE. In this patch, a
protective check is added in order not to fail on NPE when restarting
the engine and a warning is added to the log asking the user to cleanup
the command_entities table.

Change-Id: Ia49bd24da914f36f09241feae8875055ea1dd27a
Bug-Url: https://bugzilla.redhat.com/1477042
Signed-off-by: Ala Hino <ahino@redhat.com>