webadmin: convert radio buttons to button groups on data tables
Some data tables have a set of radio buttons in the header that allows
selecting the base type of list data and table columns. Example is the
Storage>Disks table or Administration>User tables.

The radio buttons are converted to bootstrap button groups so they
look better with PatternFly styles and won't get cut off.

Change-Id: I7560ec621e1b93117bb74de2604fa8ac4c5665e9
Signed-off-by: Scott J Dickerson <sdickers@redhat.com>