core: always clone on import from cfme
When importing a VM not as a new entity (i.e., not cloning it), we preserve
its MAC address. This is also the case  when the mac address is out of the
MACs pool's range, which leads to unplugged NIC(s) for the imported VM.
As a temporary workaround, we change the import from external URL (i.e.,
only from external provider by CFME at the moment) to always clone the VM,
also when the VM doesn't exist in the system. This way, the MAC addresses
are reallocated during the import process.

Change-Id: I709a0dee90eeade0f7715e3be3242a6ce15e1f65
Bug-Url: https://bugzilla.redhat.com/1437152
Signed-off-by: Arik Hadas <ahadas@redhat.com>