backend: replace network name with vdsm name when needed
In a scenario where a network with a long name is removed or detached
from a cluster, detaching the network (which is now unmanaged) from
hosts will be impossible as mapping network name to vdsm name is no
longer possible.

To remedy this, on network removal, replace network_name in
vds_interface from network name to vdsm name in each interface
where the removed network is attached to in the datacenter or cluster.

Change-Id: I93fca4940deae3bff497c85fcaf46e1a87f07c02
Bug-Id: https://bugzilla.redhat.com/show_bug.cgi?id=1478790
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>