engine: Pass plugin type for each external network allocate request.
Plugin type must be passed to vdsm whenever
a port is allocated. The plugin type can be used
to determine which provider driver will be executed,
and allow to have multiple drivers installed.
Currently the drivers have no way of knowing
which should be handling the nic.
This will also fix the problem where the
openstacknet driver can not be installed along
any of the other external network providers.
The openstack driver is checking for plugin
type before it decides to run or not to run.
With no plugin type, the openstacknet hooks
were failing, even if the nic belonged to
a different type of provider.

Bug-Url: https://bugzilla.redhat.com/1478823

Change-Id: I4f76d4f4e4e087bc88c9412705c168d4e1fa08c7
Signed-off-by: Marcin Mirecki <mmirecki@redhat.com>