webadmin: Automatic setting of IO threads for a 'High performance' VM
Automatic enabling of the "IO Threads" configuration propery for a new/edited
'High performance' VM/Template/Pool.

In addition, removing the "Num Of IO Threads" text field from the
UI dialogs of VM/Template/Pool/Instance Type while setting the num of IO
threads value to be always 1. So that in case IO Threads is enabled, only
one IO thread is always created, since it is the best configuration as of
current performance testings.

Change-Id: Ib32596c590f57e9045312dd2da22bf66c24e7cfd
Bug-Url: https://bugzilla.redhat.com/1457239
Signed-off-by: Sharon Gratch <sgratch@redhat.com>