core: write vm custom properties to engine xml
Write the custom properties that are defined for a VM to the metadata section
in the engine xml. Specifically, this solves the issue of the vfio-mdev hook
not being called because of the lack of the mdev_type property.

An example of how the metadata section looks like with custom properties:
<domain type="kvm" xmlns:ovirt-vm="http://ovirt.org/vm/1.0">
  ... skipped ...
  <metadata>
    <ovirt-vm:vm>
      <ovirt-vm:version type="float">4.2</ovirt-vm:version>
      <ovirt-vm:custom>
        <ovirt-vm:mdev_type>arik</ovirt-vm:mdev_type>
      </ovirt-vm:custom>
    </ovirt-vm:vm>
  </metadata>
</domain>

Change-Id: I0b3a78e37bcdfe095b22f9b783260edd523bd06c
Bug-Url: https://bugzilla.redhat.com/1482033
Signed-off-by: Arik Hadas <ahadas@redhat.com>