webadmin: VmSnapshotListModel PREVIEW snapshot checking
In VmSnapshotListModel#setItems, every iteration of the loop checks
if there's a snapshot which has the PREVIEW type. However, as the
snapshots do not change throughout the loop, it could be checked only
once, before the loop. This would both [arguably] clean up the code,
but also provide a slight performance enhancement.

Change-Id: I7aa3c21f4e5af89abfa0fb5413510785c73c1d5f
Signed-off-by: Allon Mureinik <amureini@redhat.com>