core: Add callback mechanism to ExportVmCommand
Add concurrent callback mechanism to ExportVmCommand.

Due to the use of concurrent callback, the tasks which
occur on CopyImageGroup are now added to the VdsmTaskIdList
instead of InternalVdsmTaskIdList.

As side effect, the audit log now works properly
in case of a VM which has no disks and exported to an
export domain.

Change-Id: Ie36772806d41ee23ff344caae3046a8188e2fa54
Bug-Url: https://bugzilla.redhat.com/1477963
Signed-off-by: Eyal Shenitzky <eshenitz@redhat.com>