core: missing injection caused NPE in event log
When trying to dismiss an error (like in the case you try to add a host
to an incompatible dc/cluster version), a NPE is thrown as a result of
a AuditLogBase instance that is created without injecting its members.

This patch handled this problem by populating an AuditLogImpl instance
instead without a need of injection.

Change-Id: Ibe53da55b969962139b43b9701f8c100a2a40306
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1480208
Signed-off-by: emesika <emesika@redhat.com>