webadmin: detect base template correctly on import template
The problem was that while the template is being imported, it is checked, if
the template has it's base template already in the system. If not, the template
has to be imported as clone.

The problem was, that when the template was a base template, it was always
enforced to be imported to the system as a clone.

Fixed by a proper check. The patch also contains one additional null check to
make sure the frontend does not throw a NPE while leaving the screen.

Change-Id: I8499d4bdc35b3ef50a723d42143691cd267ceb80
Bug-Url: https://bugzilla.redhat.com/1485804
Signed-off-by: Tomas Jelinek <tjelinek@redhat.com>