core: Remove GuidUtils
GuidUtils has a single static method, getGuidListFromString, while
Guid had all the other static methods to convert Strings to Guids.
This patch moves the method to the Guid class and renames it to
createGuidListFromString to follow the conventions of the Guid class.
Similarly, GuidUtilsTest was removed, and its tests moved to
GuidTest.

Change-Id: I921190b809b403a62baa652c313f4db13010e2f9
Signed-off-by: Allon Mureinik <amureini@redhat.com>