core: remove MultiValueMapUtils
MultiValueMapUtils was a helper class introduced back in oVirt 3.1
(commit fc441f778a4) as a mechanism to easily manage maps where the
values as collections (i.e., Map<T, S extends Collection<R>>).

With Java 8's capabilities this helper is mostly redundant, and
continuing to maintain it just causes the code to be less readable
and harder to maintain.

This patch removes this outdated helper and replaces it with one of
two patterns.
- Where it was used to convert a Collection to a Map in a
  straight-forward manner, the collection was streamed and then
  collected with Collectors.groupingBy.
- Where a more intricate logic was required, the original logic was
  kept but the call to MultiValueMapUtils.appToMap was replaced with
  a computeIfAbsent(key, k -> new ArrayList<>()) call, as per the
  recommendation in java.util.Map's documentation [1].

[1] https://docs.oracle.com/javase/8/docs/api/java/util/Map.html#computeIfAbsent-K-java.util.function.Function-

Change-Id: I6ba83b6ca99ad2a365115903d172dafe4627d328
Signed-off-by: Allon Mureinik <amureini@redhat.com>