core: Don't add memory disk if storage does not exist.
As part of the process of importing unregistered entities,
the engine tries to add any memory volumes from snapshots.
Once a storage domain does not exists for this memory volume,
the volume should not be added as a disk and a proper error
should be logged.
The VM should still be imported successfully.

Change-Id: I419a65e85086beb742de199e38dd67d04b58bc98
Bug-Url: https://bugzilla.redhat.com/1478283
Signed-off-by: Maor Lipchuk <mlipchuk@redhat.com>