packaging: setup: Add a dummy ovirt-engine-checks/base
engine-setup now requires the plugin group ovirt-engine-checks, but we
only package stuff for it in the ovirt-engine plugins subpackage.
This makes engine-setup fail if that subpackage is not installed, e.g.
on a separate DWH machine.

Add a dummy "ovirt-engine-checks/base/core" directory under the "base"
subpackage, so that "ovirt-engine-checks" always exists.

Bug-Url: https://bugzilla.redhat.com/1484961
Change-Id: I094ff482fd8821c757da2de98e78c07d7e02b63a
Signed-off-by: Yedidyah Bar David <didi@redhat.com>