core: Custom Id is not persisted to the database
When creating a audit log record, there is an option to assign a
custom id string to the record, this is used for example when dealing
 with interfaces where the custom id value gets the interface name
in order to handle event flooding correctly.
However, the custom id value is used when generating the key for the
audit log cache but since it is not persisted to the database, each
record in the cache that has a non-empty custom id value will not be
cleaned from the cache since the key for searching this record will
have always an empty custom id value.
This behavior may lead to cache flooding and memory leak.

Change-Id: I4867253abd8ef35dbab6ecd28cc2aa9a8bae15aa
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1483889
Signed-off-by: emesika <emesika@redhat.com>