backend: change exception log level to debug
When the description of an unregistered disk is not in a JSON format, an
exception is thrown by
MetadataDiskDescriptionHandler::enrichDiskByJsonDescription.
This very long exception with its traceback is logged by
GetUnregisteredDiskQuery::executeQueryCommand in warning level, and
pollutes the log with unnecessary information for the average user.
Therefore, this log level was changed to debug, and a shorter and more
informative warning log was added instead.

Change-Id: Iee539cd3b45a975f08af77b571ab615193f5cab4
Bug-Url: https://bugzilla.redhat.com/1486293
Signed-off-by: Idan Shaby <ishaby@redhat.com>