search: resolve ambiguous search parameter
The problem was that from VMs search you can not search for Hosts
attributes (i.e. "Vms:Hosts.<some attribute>"

The reason for that is that both "Hosts" and "host" are available as
search parameters. The search engine is case insensitive and therefor
since it is processing the search string character-by-character after
analysing the entity type (i.e what appears as "<entity>:" , "vms:" for
example) it get the "host" as a token when you are writing "Hosts" and
marks the "s" at the end of the "Hosts" string as an error.

The solution for that is to change the "host" parameter to any other
string which is not the prefix of "Hosts" , this patch changed "host" to
"on_host" but it can be any other value.

Change-Id: Ic28a8e0cbadd0e5f959fb0d9d7871e6d5f1cb388
Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1415751
Signed-off-by: emesika <emesika@redhat.com>