core: Move RunVmCommand backup domain validation to MSDV
Move the validation that none of the domains the VM is using are a
backup domain to MultipleStorageDomainValidator so the logic is not
reimplemented.

The tests were removed from RunVmCommand as the logic was moved to
MultipleStorageDomainValidator,(as there's no point in testing
passthrough methods), and new tests were added to
MultipleStorageDomainValidatorTest.

Change-Id: I692371e1dcfd3635e840a82015f17deb650e5601
Signed-off-by: Allon Mureinik <amureini@redhat.com>