webadmin: Expand width of new QoS popup
The existing width of the new Storage QoS popup is too narrow to
accommodate the width of the fr_FR and es_ES translations. This patch
expands the width of the dialog so that it fits properly.

Screenshots:
es_ES - http://imgur.com/a/MkaRD
fr_FR - http://imgur.com/a/ce0t9

Change-Id: Idc33ac9d7db6c79f5ab88474d2d9aca288eef4dd
Bug-Url: https://bugzilla.redhat.com/1417904
Signed-off-by: Phillip Bailey <phbailey@redhat.com>