webadmin: Further expand tab bar width in Cluster popup
The previous patch [1] to expand the tab bar width in the new/edit Cluster
popup didn't increase it enough to accommodate the width of the
ja_JP translation of the "Scheduling Policy" side tab label when the label
is selected in Firefox. This patch fixes that.

[1] https://gerrit.ovirt.org/#/c/81240/

Screenshot: http://imgur.com/a/pmwpi

Change-Id: I89c22644bdd0389422f8aeaa8b3a758bbb353419
Bug-Url: https://bugzilla.redhat.com/1418165
Signed-off-by: Phillip Bailey <phbailey@redhat.com>