webadmin: remove listener from VM name in VmImportGeneralModel
Currently, when changing the VM name in the import VM dialog,
the new name will persist even when clicking cancel (until moving
to another tab). This is caused by listener listening to changes
on the VM and persisting them.

To fix this, the listener was removed. There is no need to listen
to changes in the VM name input field since we only need to
change it when "OK" is clicked.

Change-Id: Ia15c223964886e7c1f2b28394243beeacb10157e
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>