webadmin: use a clone for ImportVmData entity
Currently, when changing the VM name in the import VM dialog,
the new name will persist even when clicking cancel (until moving
to another tab). This is caused by listener listening to changes
on the VM and persisting them.

To fix this, a clone of the VM object is used for the entity.

Change-Id: I312b66486b77ac114da4db3dd942a89b38425b16
Bug-Url: https://bugzilla.redhat.com/1449515
Signed-off-by: Benny Zlotnik <bzlotnik@redhat.com>