core: update the required fields when vm disappears
We used to avoid updating the VmDynamic data of a VM that has been
disappeared, i.e., is not being reported by the host we expect the
VM to run on, so we won't override its exit status, exit message
and exit reason. However, we must ensure that we update other
fields, like run_on_vds.

Change-Id: I0727bc36e4d6b19a6ad920d08c34369b6cdb60c9
Bug-Url: https://bugzilla.redhat.com/1487728
Signed-off-by: Arik Hadas <ahadas@redhat.com>