bll: Optimization - don't connect same targets in RunVM
Motivation:
Connecting iscsi targets, when lots of them need to be connected to a
host, may be a long operation, and may stall RunVm. Given a VM to run
with different disks that all needs the same target will cause
unnecessary repeated call to connect it.

Modification:
Memoize the targets of the lun we need to connect. If they were already
called, in this scope, don't try connect again.

Result:
A call to connect a target will be made exactly once if it was not made
already, in the scope of running VM.

Out-of-Scope: We don't handle 2 parallel RunVm with identical targets.

Change-Id: I4f20a7d799fe1bd433be36542bced78834c510bd
Bug-Url: https://bugzilla.redhat.com/1468439
Signed-off-by: Roy Golan <rgolan@redhat.com>